Ejercicio1 Pila

+++ src/main/java/com/victorlopez/tema08/pilasColas/Ejercicio1/Pila.java
@@ -3,7 +3,7 @@ package com.victorlopez.tema08.pilasColas.Ejercicio1;
 import java.util.Arrays;
 
 public class Pila {
-    private int[] pila;
+    private double[] pila;
     private int nElementos;
     private int maxArray;
 
@@ -13,7 +13,7 @@ public class Pila {
     public Pila() {
         maxArray = 10;
         nElementos = 0;
-        pila = new int[maxArray];
+        pila = new double[maxArray];
     }
 
     /**
@@ -33,8 +33,8 @@ public class Pila {
      * Método que devuelve el último valor del array
      * @return devolvemos el último valor del array
      */
-    public int top(){
-        int aux = Integer.MIN_VALUE;
+    public double top(){
+        double aux = Double.MIN_VALUE;
         if(!empty()){
             aux = pila[nElementos -1];
         }
@@ -47,7 +47,7 @@ public class Pila {
      */
     public Pila(int maxArray) {
         this.maxArray = maxArray;
-        pila = new int[maxArray];
+        pila = new double[maxArray];
         nElementos = 0;
     }
 
@@ -55,7 +55,7 @@ public class Pila {
      * Método para insertar un elemento en la pila
      * @param n entero a insertar
      */
-    public void push(int n){
+    public void push(double n){
         if (nElementos == pila.length){
             pila = duplicarArray(pila);
         }
@@ -68,9 +68,9 @@ public class Pila {
      * @param pila array a duplicar
      * @return devuelve el array duplicado
      */
-    private int[] duplicarArray(int[] pila){
+    private double[] duplicarArray(double[] pila){
         maxArray = maxArray * 2;
-        int[] aux = new int[maxArray];
+        double[] aux = new double[maxArray];
         for (int i = 0; i < pila.length; i++){
             aux[i] = pila[i];
         }
@@ -97,8 +97,8 @@ public class Pila {
      * Método para mostrar el último elemento del vector, además, lo borramos
      * @return devolvemos el último elemento
      */
-    public int pop(){
-        int aux = Integer.MIN_VALUE;
+    public double pop(){
+        double aux = Integer.MIN_VALUE;
         if(!empty()){
             aux = pila[nElementos -1];
             nElementos--;
